backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cattrs==24.1.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
kerykeion==4.26.2
MarkupSafe==3.0.2
openai==1.77.0
packaging==25.0
platformdirs==4.3.7
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pyswisseph==2.10.3.2
python-dotenv==1.1.0
pytz==2024.2
requests==2.32.3
requests-cache==1.2.1
scour==0.38.2
simple-ascii-tables==1.0.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
url-normalize==2.2.1
urllib3==2.4.0
Werkzeug==3.1.3

Flask-SQLAlchemy


pytest
pytest-flask


Flask-Migrate


pytest-cov

